Fixed serializer tests
